+++ src/compliance-suite.ts
@@ -3,8 +3,7 @@ import path from "node:path";
 import { Config } from "@interpreter/config/config";
 import { ColorManager } from "@interpreter/config/managers/color/manager";
 import { Interpreter, type InterpreterResult } from "@interpreter/interpreter";
-import { Lexer } from "@interpreter/lexer";
-import { Parser } from "@interpreter/parser";
+import { parseExpression } from "@interpreter/parser";
 import { BaseSymbolType, ColorSymbol, NumberWithUnitSymbol } from "@interpreter/symbols";
 import { groupBy } from "./interpreter/utils/coll";
 import { isArray, isObject } from "./interpreter/utils/type";
@@ -133,9 +132,7 @@ function transformContextToSymbols(obj: any, config: Config): any {
 }
 
 const runTest = (test: TestCase): { interpreter: Interpreter; result: InterpreterResult } => {
-  const lexer = new Lexer(test.input);
-  const parser = new Parser(lexer);
-  const ast = parser.parse(test.inline);
+  const { ast } = parseExpression(test.input, { mode: test.inline ? "inline" : "script" });
 
   let config: Config;
   if (test.schemas && test.schemas.length > 0) {